incorporate fastboot visit into the indexing
