+++ packages/runtime-common/loader.ts
@@ -4,11 +4,22 @@ import { Deferred } from "./deferred";
 import { trimExecutableExtension } from "./index";
 import { RealmPaths } from "./paths";
 import { CardError } from "./error";
+import { formatRFC7231 } from "date-fns";
 
 // this represents a URL that has already been resolved to aid in documenting
 // when resolution has already been performed
 export interface ResolvedURL extends URL {
   _isResolved: undefined;
+}
+
+function isResolvedURL(url: URL | ResolvedURL): url is ResolvedURL {
+  return "_isResolved" in url;
+}
+
+function makeResolvedURL(unresolvedURL: URL | string): ResolvedURL {
+  let resolvedURL = new URL(unresolvedURL) as ResolvedURL;
+  resolvedURL._isResolved = undefined;
+  return resolvedURL;
 }
 
 type RegisteredModule = {
@@ -67,6 +78,10 @@ export class Loader {
     { module: string; name: string }
   >();
   private consumptionCache = new WeakMap<object, string[]>();
+  private responseCache = new Map<
+    string,
+    { body: string; lastModified: number }
+  >();
 
   constructor() {}
 
@@ -162,6 +177,10 @@ export class Loader {
       moduleIdentifier,
       this.createModuleProxy(module, moduleIdentifier)
     );
+  }
+
+  setJSONAPIResponseBody(url: URL, body: string) {
+    this.responseCache.set(url.href, { body, lastModified: Date.now() });
   }
 
   async getConsumedModules(
@@ -269,35 +288,54 @@ export class Loader {
     urlOrRequest: string | URL | Request,
     init?: RequestInit
   ): Promise<Response> {
+    let requestURL =
+      urlOrRequest instanceof Request
+        ? urlOrRequest.url
+        : typeof urlOrRequest === "string"
+        ? urlOrRequest
+        : urlOrRequest.href;
+    let cachedResponse = this.responseCache.get(requestURL);
+    if (cachedResponse != null) {
+      return makeJSONAPIResponse(
+        cachedResponse.body,
+        cachedResponse.lastModified
+      );
+    }
+
     if (urlOrRequest instanceof Request) {
       for (let [url, handle] of this.urlHandlers) {
         let path = new RealmPaths(new URL(url));
-        if (path.inRealm(new URL(urlOrRequest.url))) {
+        if (path.inRealm(new URL(requestURL))) {
           let request = urlOrRequest as MaybeLocalRequest;
           request.isLocal = true;
           return await handle(request);
         }
       }
-      let request = new Request(this.resolve(urlOrRequest.url).href, {
+      let request = new Request(this.resolve(requestURL).href, {
         method: urlOrRequest.method,
         headers: urlOrRequest.headers,
         body: urlOrRequest.body,
       });
       return fetch(request);
     } else {
+      let unresolvedURL =
+        typeof urlOrRequest === "string"
+          ? new URL(urlOrRequest)
+          : isResolvedURL(urlOrRequest)
+          ? this.reverseResolution(urlOrRequest)
+          : urlOrRequest;
       for (let [url, handle] of this.urlHandlers) {
         let path = new RealmPaths(new URL(url));
-        if (path.inRealm(new URL(urlOrRequest))) {
+        if (path.inRealm(unresolvedURL)) {
           let request = new Request(
-            typeof urlOrRequest === "string" ? urlOrRequest : urlOrRequest.href,
+            unresolvedURL.href,
             init
           ) as MaybeLocalRequest;
           request.isLocal = true;
           return await handle(request);
         }
       }
-      let resolvedURL = this.resolve(urlOrRequest);
-      return fetch(resolvedURL.href, init);
+      return fetch(this.resolve(unresolvedURL).href, init);
     }
   }
 
@@ -306,10 +344,10 @@ export class Loader {
     for (let [sourceURL, to] of this.urlMappings) {
       let sourcePath = new RealmPaths(new URL(sourceURL));
       if (sourcePath.inRealm(absoluteURL)) {
-        return new URL(sourcePath.local(absoluteURL), to) as ResolvedURL;
+        return makeResolvedURL(new URL(sourcePath.local(absoluteURL), to));
       }
     }
-    return absoluteURL as ResolvedURL;
+    return makeResolvedURL(absoluteURL);
   }
 
   reverseResolution(
@@ -441,7 +479,7 @@ export class Loader {
     await Promise.all(
       dependencyList!.map(async (depId) => {
         if (depId !== "exports" && depId !== "__import_meta__") {
-          return await this.fetchModule(new URL(depId) as ResolvedURL, [
+          return await this.fetchModule(makeResolvedURL(depId), [
             ...stack,
             moduleIdentifier,
           ]);
@@ -552,4 +590,14 @@ export class Loader {
 
 function assertNever(value: never) {
   throw new Error(`should never happen ${value}`);
+}
+
+function makeJSONAPIResponse(body: string, lastModified: number): Response {
+  return new Response(body, {
+    status: 200,
+    headers: {
+      "last-modified": formatRFC7231(lastModified),
+      "content-type": "application/vnd.api+json",
+    },
+  });
 }
